requirements.txt
django-3.0.8
djangorestframework-3.11.0
psycopg2-2.8.5
pillow-7.2.0
